Created dlna specific object model
Somehow I've ended up with different line endings than the previous
commit so the diff on Plugin.cs is practically the entire file, but
these changes are too good to throw awy in order to fix that...
Hanselmans recent post is very timely:
http://www.hanselman.com/blog/YoureJustAnotherCarriageReturnLineFeedInTheWall.aspx

Actual changes:
Created a dlna specific object model to wrap the mediabrowser model.
It's only functional for Video and even then really only for the way
video is used on the Xbox360. Music is missing at the moment.
Utilised the new object model when Search and BrowseDirect is called.
Pulled in the MoreLinq nuget package.

The next commit will be a bit of a cleanup, with lots of commented out
code removed and lots of uncalled functions removed as well.