Added comments about Platinum Properties
I found the didl-lite xsd which has decent descriptions for some of its
elements and attributes. A lot of these map directly to Platinum
Properties so I copied a lot of the descriptions above where the
properties are being utilised. Ideally this information would be in
triple slash comments inside the Platinum Library but alas it is not.