Got image requests going through the api
Had to pass the actual MB item around instead of just the id because
there are possilby mulitple images available and the api barfs if there
are no images available. This meant a fair bit of refactoring but its
better for it.
Tweaked the  platinum ObjectClass for epsiodes to be
object.item.videoItem.videoBroadcast instead of just
object.item.videoItem - I think this will get removed once I test it on
the Xbox360 and discover it doesn't work as expected.
Added well know 'Music Genre' and 'Video Series' containers and items.