Recompiled Platinum.Managed in V90 Release Mode
Recompiled Platinum.Managed on a machine with Visual Studio 2008 on it.
Still compiled in VS2010 but used Platform Toolset V90 (vs2008)
compiler, which means its compiled against .Net 2.0 not 4.0.
Most importantly this compile is definitely in Release Mode, so should
have a dependency on msvcr100.dll and not msvcr100d.dll (the d at the
end is for Debug, and only gets installed on development machines).