Added well known All Playlists container - 13
Added well known container Playlist and its well know children "All
Playlists" and "Playlists Folder". Both of these are just returning
empty collections for their children, but it means that when WMP
requests well known container 13 "All Playlists", we don't have to
recursively load the entire data model in order to search for it.
Also removed unrequired test code that was setting xml string to a local
variable to assist with debugging.