Got browsing working on PS3
Got browsing working on the PS3 all the way down to an item - hitting
the item breaks the API which breaks the server, but still, its a start.
Had to only return the single item of metadata being requested and jsut
as importantly, had to report the number of children container objects
have - I'd neglected to set that property which meant it was assuming
all containers had no children.