New version of Platinum
New version of Platinum (code released today), new Platinum.Managed.dll
was compiled with the new code. Its still .net 2.0 compiled in Release
mode.
And I finally got Git to add the important directories - the issue was
that we've got build directories set to Ignore, usually a fine idea
except when Plutinosoft decided to put all the important stuff in a
directory under a directory called 'build'.
So while much of this checkin is reporting as 'new' it's really an
update, its just that we could never get the files to detect previously.