Started converting model to parent pointers
Started converting the model over to mainting a parent object rather
than just a parent id. The long term plan is to be able to build a
navigation path, like breadcrumbs, from any item so that we can navigate
directly to a given item rather than having to search for it.