Implemented new Path as ID instead of itemID
Implemented a new way of handing out the ID for dlna objects, the ID
handed out is now made up of the path required to navigate back to that
item. This saves having to search all containers for a given item. It
will hopefully speed up navigation on Marks TV as well as PS3 and other
clients. Will have no real effect on WMP due to the way it is
implemented.