Allow clients to call artwork uri with extension
Addd some temp code that gives clients the client a uri with a .jpg and
a .png extension. This is to test how many clients require a file
extension on the uri in order to diplay artwork.
The other api uri's are still there, but most clients will only call the
first one given, so they will be ignored for the most part.