Implemented basic upnp:class search
Implemented basic derived from and equals upnp:class search. It will
make WMP and Xbox360 Music App behave more nicely when it comes to music
items being displayed in the video section and vice-versa.
Still got issues with items being doubled up due to having different
paths to the same item, this is an ID issue and has nothing to do with
Search.