Changed Id/Path again to stop duplicate items
Put the dlna ID back to being a single ID rather than a path of IDs
because the the same video or audio item always has multiple paths (All
Music/Music Folder/Music Genre/etc) and using the path as ID made it
seem like they were different items. Now implemented a Map from ID to
Parent IDs which can be walked so that we can still navigate to an item
fairly directly.