Samsung images working using customised ProtoInfo
Added an extra contuctor to the Plainum.Manged ProtocolInfo class so
that we can create the object with the actaul protocol string in the
same manner than unmanaged developerrs can do. It's a part of the public
interface that missing in the Managed Version.
Used this new ability to provide thumnail resources for all video and
music (technically shouldn't be required for music but it doesn't hurt
to put them there). This change got us thumbnails on a Samsung Tv.