Thumbnails working on PS3, Icons attempted
Got thumbnails working on the PS3 by supplying the resolution in the
Thumbnail's Resource info - PS3 requires a PNG_TN (PNG ThumNail) or a
JPEG_TN and it requires it to be reported to be 160x160, which is what a
PNG_TN is supposed to be anyway - however, if you serve it a 160x160 it
looks jagged and horrible, so I'm serving it an 800x800 and it looks
great.
Made some tweaks to serve Icon urls for all items as well, although I
don't think any clients ever look for the icon url.
Changes to video info thats served to force the xbox to request a wmv
trascode - which works.