Add an interface for Comic metadata from files
A comic stored in the comic book archive supports various file formats
for metadata. Jellyfin currently supports reading metadata stored
in the ComicInfo format when saved as an external file.

This change adds an interface and a proxy implementation to allow
Jellyfin to check multiple metadata formats and save locations for the
same comic.

Co-authored-by: Roland Helmich <helmich.roland@gmail.com>