Add ComicBookInfo metadata Providers
The ComicBookInfo format stores metadata within the zip archives
comment. As System.IO.Compression does not expose the zip archives
comment, SharpCompress is used instead.

Reads in all fields the ComicBookInfo defines and saves as many as
possible in Jellyfin. Additional fields can be saved once support
for these is added in the Jellyfin Server project.