Add development setup for Visual Studio Code (#64)
* Add development setup for Visual Studio Code

The implementation of the development setup is based upon the
implementation in the plugin template repository with addition of proper
linux support.

The ${env:XDG_DATA_HOME} variable is not used, as this variable resolves
to the current working directory within Visual Studio Code's development
container. By manually specifying the file path, it is possible to use
Visual Studio Code's development container.

The "mkdir" command uses the "-p" flag. This flag creates the parent
directories, as well. Should they already exists, the command moves
down to the next folder to create without throwing an error.

The "-r" parameter of the "cp" is required to recursively copy all
files and directories within the "publish" directory.

* Add required extensions to Visual Studio Code